add auth for POST PATCH DELETE

+++ src/controllers/breads-controller.ts
@@ -9,36 +9,38 @@ class BreadsController {
     const breadID = req.params.breadID;
 
     if (breadID) {
-      BreadModel.findById(breadID)
-        .then((bread: Bread) => {
-          if (bread) {
-            res.status(200).json(bread);
-          } else {
-            res.status(404).json({ error: "Object doesn't exist.", breadID });
-          }
-        })
-        .catch((err: CallbackError) =>
-          res.status(500).json({ error: `${err}` })
-        );
+      if (breadID.length === 24) {
+        BreadModel.findById(breadID)
+          .then((bread: Bread) => {
+            if (bread) {
+              res.status(200).json(bread);
+            } else {
+              res.status(404).json({ error: "Object doesn't exist.", breadID });
+            }
+          })
+          .catch((err: CallbackError) => res.status(500).json({ error: err }));
+      } else {
+        res.status(404).json({ error: "Incorrect id format." });
+      }
     } else {
       BreadModel.find()
         .then((breads: Bread[]) => res.status(200).json(breads))
-        .catch((err: CallbackError) =>
-          res.status(500).json({ error: `${err}` })
-        );
+        .catch((err: CallbackError) => res.status(500).json({ error: err }));
     }
   };
 
   postHandler: RequestHandler = (req, res, next) => {
-    const bread = new BreadModel(req.body);
+    const bread = new BreadModel({ ...req.body, creator: req.params.userID });
     const errors = validationResult(req);
 
     if (errors.isEmpty()) {
       bread
         .save()
-        .then((bread) => res.status(201).json(bread))
+        .then((bread) =>
+          res.status(201).json({ message: "Created successfully.", bread })
+        )
         .catch((err: CallbackError) => {
-          res.status(500).json({ error: `${err}` });
+          res.status(500).json({ error: err });
         });
     } else {
       return res.status(422).json({ errors: errors.array() });
@@ -50,13 +52,38 @@ class BreadsController {
     const errors = validationResult(req);
 
     if (errors.isEmpty()) {
-      BreadModel.findOneAndUpdate(breadID, req.body, {
-        useFindAndModify: false,
-      })
-        .then(() => res.status(201).json({ ...breadID, ...req.body }))
-        .catch((err: CallbackError) => {
-          res.status(500).json({ error: `${err}` });
+      if (breadID._id.length === 24) {
+        BreadModel.findOne(breadID).then((bread: Bread) => {
+          if (bread) {
+            if (bread.creator === req.params.userID) {
+              BreadModel.findOneAndUpdate(breadID, req.body, {
+                useFindAndModify: false,
+              })
+                .then(() => {
+                  res.status(201).json({
+                    message: "Updated successfully.",
+                    ...breadID,
+                    ...req.body,
+                  });
+                })
+                .catch((err: CallbackError) => {
+                  res.status(500).json({ error: err });
+                });
+            } else {
+              res.status(403).json({
+                error:
+                  "You are not allowed to modify objects of other creators.",
+              });
+            }
+          } else {
+            res
+              .status(404)
+              .json({ error: "Object doesn't exist.", ...breadID });
+          }
         });
+      } else {
+        res.status(404).json({ error: "Incorrect id format." });
+      }
     } else {
       return res.status(422).json({ errors: errors.array() });
     }
@@ -65,19 +92,31 @@ class BreadsController {
   deleteHandler: RequestHandler = (req, res, next) => {
     const breadID = { _id: req.params.breadID };
 
-    BreadModel.findByIdAndDelete(breadID)
-      .then((bread: Bread) => {
+    if (breadID._id.length === 24) {
+      BreadModel.findOne(breadID).then((bread: Bread) => {
         if (bread) {
-          res
-            .status(200)
-            .json({ message: "Deleted successfully.", ...breadID });
+          if (bread.creator === req.params.userID) {
+            BreadModel.findByIdAndDelete(breadID)
+              .then(() => {
+                res
+                  .status(200)
+                  .json({ message: "Deleted successfully.", ...breadID });
+              })
+              .catch((err: CallbackError) => {
+                res.status(500).json({ error: err });
+              });
+          } else {
+            res.status(403).json({
+              error: "You are not allowed to modify objects of other creators.",
+            });
+          }
         } else {
           res.status(404).json({ error: "Object doesn't exist.", ...breadID });
         }
-      })
-      .catch((err: CallbackError) => {
-        res.status(500).json({ error: `${err}` });
       });
+    } else {
+      res.status(404).json({ error: "Incorrect id format." });
+    }
   };
 }
